requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
Django==4.2.5
django-cors-headers==4.2.0
django-countries==7.5.1
django-crispy-forms==2.0
django-jazzmin==2.6.0
django-ninja==0.22.2
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==10.0.1
pydantic==1.10.12
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
